+++ loaders/binary.js
@@ -1,4 +1,0 @@
-export default async function (path) {
-  const content = await Deno.readFile(path);
-  return { content };
-}