+++ loaders/json.js
@@ -1,4 +1,0 @@
-export default async function (path) {
-  const content = await Deno.readTextFile(path);
-  return JSON.parse(content);
-}