+++ loaders/json.ts
@@ -1,0 +1,9 @@
+import { Data } from "../types.ts";
+
+/**
+ * Load and parse a JSON file.
+ */
+export default async function (path: string): Promise<Data> {
+  const content = await Deno.readTextFile(path);
+  return JSON.parse(content);
+}