+++ plugins/eta.js
@@ -1,6 +1,6 @@
 import * as eta from "../deps/eta.ts";
 import Eta from "../engines/eta.ts";
-import loader from "../loaders/text.js";
+import loader from "../loaders/text.ts";
 import { merge } from "../utils.ts";
 
 // Default options